requirements.txt
sys
PyQt5
sweetviz
pandas
sklearn.model_selection
lazypredict